Split up the test file
